Implemented "clear" button functionality
This feature completes all of the basic requirements for the project.
Next goals are to complete the functionalities of the rest of the "action"
buttons , add support for decimal numbers, add support for keyboard,
and fix display to properly hide overflow (currently overflow is hidden
but display text does not stay aligned to right... maybe this is a feature
we want to keep).